[fixed] Lwjgl now also handles the escape key. BECAUSE I CAN! [fixed] cosmetics [added] Rectangle has a convenience method...

+++ gdx/src/com/badlogic/gdx/math/Rectangle.java
@@ -143,7 +143,7 @@ public class Rectangle implements Serializable {
 					y + height < rectangle.y);
 	}
 
-	public void set (int x, int y, int width, int height) {
+	public void set (float x, float y, float width, float height) {
 		this.x = x;
 		this.y = y;
 		this.width = width;
